refactor: remove duplicate FrontmatterService from services/
Closes #150

FrontmatterService existed in two locations (303 lines each):
- packages/core/src/services/FrontmatterService.ts (REMOVED)
- packages/core/src/utilities/FrontmatterService.ts (KEPT)

Both files were identical. Utilities is the correct location for
helper functions. Core index.ts already exported from utilities.

No functionality changed - pure cleanup.
