+++ packages/core/src/services/TaskStatusService.ts
@@ -1,5 +1,5 @@
 import { TFile, Vault } from "obsidian";
-import { FrontmatterService } from "./FrontmatterService";
+import { FrontmatterService } from "../utilities/FrontmatterService";
 import { DateFormatter } from "../utilities/DateFormatter";
 import { EffortStatusWorkflow } from "./EffortStatusWorkflow";
 import { StatusTimestampService } from "./StatusTimestampService";
